refactor: multiset package structure
