+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/utils/CountsMap.kt
@@ -1,8 +1,8 @@
 package xyz.lbres.kotlinutils.set.multiset.utils
 
 import xyz.lbres.kotlinutils.internal.constants.Suppressions
-import xyz.lbres.kotlinutils.set.multiset.const.ConstMultiSetImpl
-import xyz.lbres.kotlinutils.set.multiset.const.ConstMutableMultiSetImpl
+import xyz.lbres.kotlinutils.set.multiset.impl.ConstMultiSetImpl
+import xyz.lbres.kotlinutils.set.multiset.impl.ConstMutableMultiSetImpl
 import xyz.lbres.kotlinutils.utils.simpleIf
 import kotlin.math.min
