+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/testutils/commonMapToSetTests.kt
@@ -3,7 +3,7 @@ package xyz.lbres.kotlinutils.set.multiset.testutils
 import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.list.copyWithoutLast
 import xyz.lbres.kotlinutils.set.multiset.MultiSet
-import xyz.lbres.kotlinutils.set.multiset.const.ConstMultiSet
+import xyz.lbres.kotlinutils.set.multiset.ConstMultiSet
 import xyz.lbres.kotlinutils.set.multiset.emptyMultiSet
 import xyz.lbres.kotlinutils.set.multiset.multiSetOf
 import xyz.lbres.kotlinutils.utils.simpleIf
