+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/utils/OperatorsTest.kt
@@ -2,7 +2,7 @@ package xyz.lbres.kotlinutils.set.multiset.utils
 
 import xyz.lbres.kotlinutils.CompList
 import xyz.lbres.kotlinutils.set.multiset.MultiSet
-import xyz.lbres.kotlinutils.set.multiset.const.ConstMultiSet
+import xyz.lbres.kotlinutils.set.multiset.ConstMultiSet
 import xyz.lbres.kotlinutils.set.multiset.emptyMultiSet
 import xyz.lbres.kotlinutils.set.multiset.multiSetOf
 import xyz.lbres.kotlinutils.set.multiset.mutableMultiSetOf
